+++ CHANGELOG.md
@@ -141,6 +141,8 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 - Deprecated `fetchMoveFunctionAbi` and `convertCallArgument`
 - Bump `aptos-client` to 1.1.0
 
+- Add permissioned signer support for FungibleAssetPermission, GasPermission, and NFTPermission using `Ed25519Account`s
+
 # 1.35.0 (2025-02-11)
 
 - Add `MultiEd25519Account` to support the legacy MultiEd25519 authentication scheme.